Changed histogram type to short to save space.

+++ src/rt/colortab.c
@@ -51,7 +51,7 @@ static CNODE	*ctree = NULL;
 				/* our color correction map */
 static BYTE	clrmap[3][256];
 				/* histogram of colors used */
-static unsigned	histo[NRED][NGRN][NBLU];
+static unsigned short	histo[NRED][NGRN][NBLU];
 				/* initial color cube boundary */
 static int	CLRCUBE[3][2] = {0,NRED,0,NGRN,0,NBLU};
